+++ src/NamingStrategy/UserIsolationNamingStrategy.php
@@ -14,28 +14,13 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
  */
 class UserIsolationNamingStrategy implements NamingStrategy
 {
-    /**
-     * @var NamingStrategy
-     */
-    private $inner;
-
-    /**
-     * @var TokenStorageInterface|null
-     */
-    private $tokenStorage;
-
-    /**
-     * @param NamingStrategy             $inner
-     * @param TokenStorageInterface|null $tokenStorage
-     */
-    public function __construct(NamingStrategy $inner, ?TokenStorageInterface $tokenStorage = null)
-    {
-        $this->inner        = $inner;
-        $this->tokenStorage = $tokenStorage;
+    public function __construct(
+        private readonly NamingStrategy         $inner,
+        private readonly ?TokenStorageInterface $tokenStorage = null,
+    ) {
     }
 
     /**
-     * @inheritDoc
      * @throws \Exception
      */
     public function createName(MutexRequest $config, Request $request): string
@@ -50,7 +35,6 @@ class UserIsolationNamingStrategy implements NamingStrategy
     }
 
     /**
-     * @return string
      * @throws \Exception
      */
     private function getToken(): string
@@ -61,6 +45,6 @@ class UserIsolationNamingStrategy implements NamingStrategy
 
         $token = $this->tokenStorage->getToken();
 
-        return $token ? md5($token->serialize()) : '';
+        return $token ? md5(serialize($token)) : '';
     }
 }